#1895 use of uuid class Signed-off-by: Florent BENOIT <fbenoit@codenvy.com>
